requirements.txt
asttokens==2.4.1
einops==0.8.0
numpy==1.22.0
opencv-python==4.8.0.74 
pillow==9.5.0
PyYAML==6.0.1
scikit-image==0.21.0
scikit-learn==1.3.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
timm==0.6.13
tqdm==4.66.4
fastapi==0.103.1
uvicorn==0.22.0
pydantic==1.10.9
torch==1.13.1
torchvision==0.14.1

